Address final declarative input review
