+++ test/shared/column-order.test.ts
@@ -1,150 +1,18 @@
 import { expect } from "@std/expect";
 import { describe, it as test } from "@std/testing/bdd";
+import { COLUMN_LAYOUTS } from "#shared/column-layout.ts";
 import type { ColumnDef, ColumnGenerators } from "#shared/column-order.ts";
 import {
-  buildDefaultTemplate,
-  COLUMN_LAYOUTS,
   getHeaderText,
   renderCells,
   renderFilteredValue,
 } from "#shared/column-order.ts";
-import {
-  EDITOR_LISTING_LAYOUT,
-  EDITOR_LISTING_TABLE_COLUMNS,
-  LISTING_TABLE_COLUMNS,
-} from "#shared/columns/listing-columns.ts";
+import { LISTING_TABLE_COLUMNS } from "#shared/columns/listing-columns.ts";
 import { escapeHtml } from "#templates/layout.tsx";
 import { setupTestEncryptionKey } from "#test-utils/env.ts";
 import { testListingWithCount } from "#test-utils/factories.ts";
 
 setupTestEncryptionKey();
-
-describe("column layout validation", () => {
-  test("derives every listing schema from the matching column definitions", () => {
-    expect([...COLUMN_LAYOUTS.listing.schema.options].sort()).toEqual(
-      Object.keys(LISTING_TABLE_COLUMNS).sort(),
-    );
-    expect([...EDITOR_LISTING_LAYOUT.columnKeys].sort()).toEqual(
-      Object.keys(EDITOR_LISTING_TABLE_COLUMNS).sort(),
-    );
-  });
-
-  test("returns null for valid template", () => {
-    expect(COLUMN_LAYOUTS.listing.validate("{{name}}, {{status}}")).toBeNull();
-  });
-
-  test("handles wonky spacing", () => {
-    expect(
-      COLUMN_LAYOUTS.listing.validate(
-        "{{ name }},{{description}},  {{ status  }}",
-      ),
-    ).toBeNull();
-  });
-
-  test("rejects unknown column (typo)", () => {
-    const error = COLUMN_LAYOUTS.listing.validate("{{name}}, {{descritpion}}");
-    expect(error).toContain("descritpion");
-    expect(error).toContain("Available columns");
-  });
-
-  test("lists the schema columns in the error message", () => {
-    const error = COLUMN_LAYOUTS.attendee.validate("{{bogus}}");
-    expect(error).toBe(
-      `Unknown column "bogus". Available columns: ${COLUMN_LAYOUTS.attendee.schema.options.join(", ")}`,
-    );
-  });
-
-  test("accepts an empty template as the default layout", () => {
-    expect(COLUMN_LAYOUTS.listing.validate("")).toBeNull();
-  });
-
-  test("rejects a non-empty template with no column tags", () => {
-    expect(COLUMN_LAYOUTS.listing.validate("name, status")).toBe(
-      "Template must include at least one column",
-    );
-  });
-
-  test("accepts templates with date filter", () => {
-    expect(
-      COLUMN_LAYOUTS.listing.validate('{{name}}, {{created | date: "%B"}}'),
-    ).toBeNull();
-  });
-
-  test("accepts templates with currency filter", () => {
-    expect(
-      COLUMN_LAYOUTS.listing.validate("{{name}}, {{price | currency}}"),
-    ).toBeNull();
-  });
-});
-
-test("keeps the attendee default column order", () => {
-  expect(COLUMN_LAYOUTS.attendee.defaultOrder).toEqual([
-    "status",
-    "date",
-    "name",
-    "listings",
-    "email",
-    "phone",
-    "address",
-    "special_instructions",
-    "answers",
-    "qty",
-    "ticket",
-    "registered",
-  ]);
-});
-
-describe("column layout parsing", () => {
-  test("returns default order when template is empty", () => {
-    const { columnKeys, filters } = COLUMN_LAYOUTS.listing.parse("");
-    expect(columnKeys).toEqual(COLUMN_LAYOUTS.listing.defaultOrder);
-    expect(filters.size).toBe(0);
-  });
-
-  test("returns columns in template order", () => {
-    const { columnKeys } = COLUMN_LAYOUTS.listing.parse("{{status}}, {{name}}");
-    expect(columnKeys).toEqual(["status", "name"]);
-  });
-
-  test("deduplicates repeated columns", () => {
-    const { columnKeys } = COLUMN_LAYOUTS.listing.parse(
-      "{{name}}, {{name}}, {{status}}",
-    );
-    expect(columnKeys).toEqual(["name", "status"]);
-  });
-
-  test("throws for an invalid template", () => {
-    expect(() => COLUMN_LAYOUTS.listing.parse("{{bogus}}")).toThrow(
-      'Unknown column "bogus"',
-    );
-  });
-
-  test("extracts filter expression for filtered column", () => {
-    const { filters } = COLUMN_LAYOUTS.listing.parse(
-      '{{name}}, {{created | date: "%B %d"}}',
-    );
-    expect(filters.get("created")).toBe('created | date: "%B %d"');
-  });
-
-  test("does not create filter entry for unfiltered column", () => {
-    const { filters } = COLUMN_LAYOUTS.listing.parse(
-      '{{name}}, {{created | date: "%B %d"}}',
-    );
-    expect(filters.has("name")).toBe(false);
-  });
-
-  test("resolves all attendee columns from default template", () => {
-    const template = buildDefaultTemplate(COLUMN_LAYOUTS.attendee.defaultOrder);
-    const { columnKeys } = COLUMN_LAYOUTS.attendee.parse(template);
-    expect(columnKeys).toEqual(COLUMN_LAYOUTS.attendee.defaultOrder);
-  });
-});
-
-describe("buildDefaultTemplate", () => {
-  test("joins column tags with a comma and space", () => {
-    expect(buildDefaultTemplate(["a", "b", "c"])).toBe("{{a}}, {{b}}, {{c}}");
-  });
-});
 
 describe("renderFilteredValue", () => {
   test("applies date filter with strftime format", () => {
